Add blueprint for user
