+++ src/templates/base.html
@@ -18,7 +18,7 @@
               <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/user/home">User</a>
+              <a class="nav-link" href="{{ url_for('users.user_home')}}">User</a>
             </li>
             {% if current_user.is_authenticated %}
             <li class="nav-item dropdown">
@@ -29,7 +29,7 @@
                 <a class="dropdown-item" href="/post/add">Add post</a>
                 <a class="dropdown-item" href="/post/view">View posts</a>
             <li class="nav-item">
-              <a class="nav-link" href="/logout">Logout</a>
+              <a class="nav-link" href="{{ url_for('users.logout')}}">Logout</a>
             </li>
             {% if current_user._id in admin_list %}
             <li class="nav-item">
@@ -38,10 +38,10 @@
             {% endif %}
             {% else %}
             <li class="nav-item">
-              <a class="nav-link" href="/login">Login</a>
+              <a class="nav-link" href="{{ url_for('users.login')}}">Login</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/signup">Signup</a>
+              <a class="nav-link" href="{{ url_for('users.signup')}}">Signup</a>
             </li>
             {% endif %}
           </ul>
