+++ src/templates/user/update.html
@@ -22,16 +22,10 @@
         <button type="submit" class="btn btn-primary">update</button>
       </form>
     </div>
-    <!-- Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
     {% else %}
-
-    <div class="container">
-      <h1 class="mt-5 mb-3">You are not authorized to update this user!</h1>
-    </div>
-
+      <div class="container">
+        <h1 class="mt-5 mb-3">You are not authorized to update this user!</h1>
+      </div>
     {% endif %}
 </body>
 </html>
