+++ src/utils/user_utils.py
@@ -2,26 +2,52 @@
 Module for user related functions
 """
 from pathlib import Path
-from flask import flash, redirect, render_template, url_for, request
-from flask_login import current_user
+from flask import flash, redirect, render_template, url_for, request, session
+from flask_login import current_user, login_user, logout_user
 
 
 class UserUtils:
     """Class for user related functions"""
 
-    def __init__(self, db, Users, user_dir):
+    def __init__(self, db, Users):
         self.db = db
         self.Users = Users
-        self.user_dir = Path(user_dir)
 
     def user_home(self):
         """User home page"""
-        return render_template(str(self.user_dir / "user_home.html"))
+        return render_template("user_home.html")
 
     def user(self, id):
         """User profile"""
         user = self.Users.query.filter_by(_id=id).first()
-        return render_template(str(self.user_dir / "user.html"), user=user)
+        return render_template("user.html", user=user)
+
+    def login(self, form):
+            """Login page"""
+            if form.validate_on_submit():
+                name = form.name.data
+                password = form.password.data
+                found_user = self.Users.query.filter_by(name=name).first()
+                if found_user:
+                    if found_user.verify_password(password):
+                        login_user(found_user)
+                        session.permanent = True
+                        session["user"] = name
+                        flash("Logged in successfully!", "info")
+                        return redirect(url_for("users.user_home"))
+                    else:
+                        flash("Invalid credentials!", "error")
+                        return redirect(url_for("users.login"))
+                flash("This user doesn't exist!", "error")
+                return redirect(url_for("users.login"))
+            return render_template("login.html", form=form)
+
+    def logout(self):
+        """Logout page"""
+        logout_user()
+        session.clear()
+        flash("Logged out successfully!", "info")
+        return redirect(url_for("users.login"))
 
     def add_user(self, form):
         """Sign up page"""
@@ -38,14 +64,14 @@ class UserUtils:
             form.password.data = ""
             if found_user:
                 flash("User already exists!")
-                return redirect(url_for("login"))
+                return redirect(url_for("users.login"))
             else:
                 usr = self.Users(name=name, email=email, password=password)
                 self.db.session.add(usr)
                 self.db.session.commit()
                 flash("User created!")
-                return redirect(url_for("login"))
-        return render_template(str(self.user_dir / "signup.html"), form=form)
+                return redirect(url_for("users.login"))
+        return render_template("signup.html", form=form)
 
     def update_user(self, id, form):
         """Update user in db"""
@@ -56,11 +82,11 @@ class UserUtils:
             try:
                 self.db.session.commit()
                 flash("User updated!", "info")
-                return render_template(str(self.user_dir / "user.html"))
+                return render_template("user_home.html")
             except:
                 flash("There was an issue updating your task", "error")
                 return render_template(
-                    str(self.user_dir / "update.html"),
+                    "update.html",
                     form=form,
                     name_to_update=name_to_update,
                 )
@@ -68,7 +94,7 @@ class UserUtils:
             form.name.data = name_to_update.name
             form.email.data = name_to_update.email
             return render_template(
-                str(self.user_dir / "update.html"),
+                "update.html",
                 form=form,
                 name_to_update=name_to_update,
                 id=id,
@@ -83,13 +109,13 @@ class UserUtils:
                 self.db.session.delete(name_to_delete)
                 self.db.session.commit()
                 flash("User deleted!", "info")
-                return redirect(url_for("user"))
+                return redirect(url_for("users.user"))
             except:
                 flash("There was an issue deleting your task", "error")
-                return redirect(url_for("user"))
+                return redirect(url_for("users.user"))
         else:
             flash("You can't delete this profile!", "error")
-            return redirect(url_for("user"))
+            return redirect(url_for("users.user"))
 
     def get_user(self, name):
         """Get user from db"""
